Update tests to validate Shortcut Name/ShortName handling
Resolves wixtoolset/issues#4227
